Migrate pricingFactory to Angular [stage-20]
