+++ src/app/purchase/purchase.module.ts
@@ -1,0 +1,13 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { PricingService } from './services/pricing.service';
+
+@NgModule({
+  declarations: [],
+  imports: [
+    CommonModule
+  ]
+})
+export class PurchaseModule {
+  static providers = [ PricingService ]
+}
